# ops/openapi-actions.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAASOA Knowledge Actions",
    "version": "0.2.0",
    "description": "Trusted retrieval and ingestion for enterprise knowledge. Authenticate with a Bearer API key."
  },
  "servers": [
    {
      "url": "https://YOUR-RAASOA-HOST"
    }
  ],
  "paths": {
    "/v1/ingest": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Ingest a document into the knowledge base",
        "description": "Upload a document (PDF, DOCX, Markdown, text, …). It is parsed, chunked, embedded, quality-checked and indexed for retrieval. Returns the quality score and any quality findings.",
        "operationId": "ingestDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingestDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/retrieve": {
      "post": {
        "tags": [
          "retrieval"
        ],
        "summary": "Search the knowledge base",
        "description": "Search trusted enterprise knowledge and return ranked, source-attributed passages with a confidence score. Use this to answer any question that depends on the organization's documents, policies, or records. Supports metadata pre-filtering (e.g. only approved documents) and per-source filtering.",
        "operationId": "searchKnowledge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_ingestDocument": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ingestDocument"
      },
...
      "ChunkHit": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title"
          },
          "chunk_type": {
            "type": "string",
            "title": "Chunk Type"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "semantic_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semantic Rank"
          },
          "lexical_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lexical Rank"
          },
          "document_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Title"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "source_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Location"
          },
          "doc_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Metadata"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "text",
          "section_title",
          "chunk_type",
          "score",
          "semantic_rank",
          "lexical_rank"
        ],
        "title": "ChunkHit"
      },
...
      "ConfidenceInfo": {
        "properties": {
          "retrieval_confidence": {
            "type": "number",
            "title": "Retrieval Confidence"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count"
          },
          "top_score": {
            "type": "number",
            "title": "Top Score"
          },
          "answerable": {
            "type": "boolean",
            "title": "Answerable"
          }
        },
        "type": "object",
        "required": [
          "retrieval_confidence",
          "source_count",
          "top_score",
          "answerable"
        ],
        "title": "ConfidenceInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IndexHit": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "predicate": {
            "type": "string",
            "title": "Predicate"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source_documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Documents",
            "default": []
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "predicate",
          "value",
          "confidence"
        ],
        "title": "IndexHit",
        "description": "Direct answer from the knowledge index (no embedding needed)."
      },
      "IngestResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status",
            "default": "auto_published"
          },
          "conflict_status": {
            "type": "string",
            "title": "Conflict Status",
            "default": "none"
          },
          "quality_findings": {
            "items": {
              "$ref": "#/components/schemas/QualityFindingSummary"
            },
            "type": "array",
            "title": "Quality Findings",
            "default": []
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "status",
          "chunk_count",
          "version",
          "embedding_model",
          "message"
        ],
        "title": "IngestResponse"
      },
...
      "QualityFindingSummary": {
        "properties": {
          "finding_type": {
            "type": "string",
            "title": "Finding Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "finding_type",
          "severity"
        ],
        "title": "QualityFindingSummary"
      },
...
      "RetrieveRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "principal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Principal Id",
            "description": "User/group ID for ACL filtering."
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type",
            "description": "Pre-filter by source type (e.g. 'sharepoint', 'jira', 'notion')."
          },
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type",
            "description": "Pre-filter by document type (e.g. 'pdf', 'policy')."
          },
          "metadata_filter": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Filter",
            "description": "Filter by frontmatter metadata. E.g. {'ampel': 'grün', 'executor': 'claude'}"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrieveRequest"
      },
      "RetrieveResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "routed_to": {
            "type": "string",
            "title": "Routed To",
            "default": "rag"
          },
          "routing_reason": {
            "type": "string",
            "title": "Routing Reason",
            "default": "default_rag"
          },
          "index_hits": {
            "items": {
              "$ref": "#/components/schemas/IndexHit"
            },
            "type": "array",
            "title": "Index Hits",
            "default": []
          },
          "structured": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructuredAnswer"
              },
              {
                "type": "null"
              }
            ]
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ChunkHit"
            },
            "type": "array",
            "title": "Results",
            "default": []
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceInfo"
          }
        },
        "type": "object",
        "required": [
          "query",
          "confidence"
        ],
        "title": "RetrieveResponse"
      },
...
      "StructuredAnswer": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          },
          "query_type": {
            "type": "string",
            "title": "Query Type"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "data",
          "query_type"
        ],
        "title": "StructuredAnswer",
        "description": "Response from structured (non-RAG) query."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "RAASOA API key as a Bearer token."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}